requirements.txt
flask
flask-login
flask-sqlalchemy
python-dotenv
twilio
google-generativeai